requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
ultralytics==8.0.200
numpy==1.24.3
opencv-python-headless==4.8.1.78
pillow==10.1.0
pandas==2.1.3
openpyxl==3.1.2
reportlab==4.0.7
supabase==2.0.3
httpx==0.25.2
aiofiles==23.2.1
torch==2.1.0
torchvision==0.16.0